レビュー対応: UserSocialType を skeleton へ移動し、アイコン定義をコンポーネント内へ集約
