+++ client/component/atom/user-social-type-select/user-social-type-select.scss
@@ -1,10 +1,6 @@
 @charset "utf-8";
 
 
-.root {
-  inline-size: 40zu;
-}
-
 .label {
   column-gap: 1zu;
   display: flex;
