+++ client/component/atom/user-social-type-select/user-social-type-select.tsx
@@ -1,9 +1,26 @@
 //
 
+import {IconDefinition} from "@fortawesome/fontawesome-svg-core";
+import {faBluesky, faDiscord, faXTwitter} from "@fortawesome/free-brands-svg-icons";
+import {faGlobe, faLink} from "@fortawesome/sharp-regular-svg-icons";
 import {ReactElement} from "react";
 import {AdditionalProps, GeneralIcon, Select, SelectOption, useTrans} from "zographia";
 import {create} from "/client/component/create";
-import {USER_SOCIAL_TYPES, USER_SOCIAL_TYPE_ICONS, UserSocialType} from "/client/constant/user-social";
+import {USER_SOCIAL_TYPES, UserSocialType} from "/server/internal/skeleton";
+
+
+/** 各リンク種別に対応するアイコン。
+ * Font Awesome に専用のブランドアイコンがない種別 (misskey, note, migdal など) には、暫定的に汎用アイコンを割り当てている (後でカスタムアイコンに差し替える予定)。*/
+const USER_SOCIAL_TYPE_ICONS = {
+  website: faGlobe,
+  x: faXTwitter,
+  bluesky: faBluesky,
+  misskey: faLink,
+  note: faLink,
+  migdal: faLink,
+  discord: faDiscord,
+  other: faLink
+} satisfies Record<UserSocialType, IconDefinition>;
 
 
 export const UserSocialTypeSelect = create(
@@ -24,19 +41,15 @@ export const UserSocialTypeSelect = create(
 
     return (
       <Select styleName="root" value={type} onSet={onSet} {...rest}>
-        {optionTypes.map((optionType) => {
-          const labelNode = (
-            <span styleName="label">
-              <GeneralIcon styleName="icon" icon={USER_SOCIAL_TYPE_ICONS[optionType]}/>
-              {trans(`label.${optionType}`)}
-            </span>
-          );
-          return (
-            <SelectOption key={optionType} value={optionType} label={labelNode}>
-              {labelNode}
-            </SelectOption>
-          );
-        })}
+        {optionTypes.map((optionType) => (
+          <SelectOption
+            key={optionType}
+            value={optionType}
+            label={<span styleName="label"><GeneralIcon styleName="icon" icon={USER_SOCIAL_TYPE_ICONS[optionType]}/>{trans(`label.${optionType}`)}</span>}
+          >
+            <span styleName="label"><GeneralIcon styleName="icon" icon={USER_SOCIAL_TYPE_ICONS[optionType]}/>{trans(`label.${optionType}`)}</span>
+          </SelectOption>
+        ))}
       </Select>
     );
